+++ src/riskmgmt/RMMakeOrders.sol
@@ -20,7 +20,7 @@ contract RMMakeOrders is RiskMgmtInterface {
 
       /// @notice Checks if the makeOrder price is within maximum allowed deviation from reference price
       /// @param orderPrice Price of Order
-      /// @param referencePrice Reference price obtained through DataFeed contract
+      /// @param referencePrice Reference price obtained through PriceFeed contract
       /// @param sellAsset Asset (as registered in Asset registrar) to be sold
       /// @param buyAsset Asset (as registered in Asset registrar) to be bought
       /// @param sellQuantity Quantity of sellAsset to be sold
@@ -45,7 +45,7 @@ contract RMMakeOrders is RiskMgmtInterface {
 
       /// @notice Checks if the takeOrder price is within maximum allowed deviation from reference price
       /// @param orderPrice Price of Order
-      /// @param referencePrice Reference price obtained through DataFeed contract
+      /// @param referencePrice Reference price obtained through PriceFeed contract
       /// @param sellAsset Asset (as registered in Asset registrar) to be sold
       /// @param buyAsset Asset (as registered in Asset registrar) to be bought
       /// @param sellQuantity Quantity of sellAsset to be sold